Build screen / gui showing 2 channels and trigs

+++ MidiIo.h
@@ -9,7 +9,7 @@
 BEGIN_MIDI_IO_NAMESPACE
 
 void init(
-    void (*onStep)(int step, void (*sendNote)(uint8_t channel, uint8_t *notes)),
+    void (*onStep)(uint8_t step, void (*sendNote)(uint8_t channel, uint8_t *notes)),
     void (*onNote)(uint8_t* chord),
     void (*onStop)(void));
